story model to story schema/added removestory logic/fixed header component
